Made the following changes to __init__.py:  - Added support for running medallion behind gunicorn for more robust multithreaded hosting.  - Added unhandled exception handler to prevent leaking internal error details.
Made the following changes to backends/mongo_backend.py:
 - Added the _type property to manifest object. This is required for filtering by type.
   e.g. manifests/?match[type]=identity
 - Fixed defect in server_discovery method where api_roots property being returned was
   not compliant with the TAXII 2.0 standard.

Made the following changes to filters/mongodb_filter.py:
 - Modified to use new "_type" property on manifest object.
 - Modified process_filter to do all filter operations in Mongodb rather than getting all
   objects in the collection back and filtering them in memory using python. This has
   massively improved performance.

Changed the format_datetime to use the existing method to convert to stix format. This
fixes a consistency issue.

Modified views/__init__.py to add a new FILTERS constant to hold the list of allowable
filters.

Added a new FILTERS import to views/manifest.py and modified the code to use that imported
filter list.

Added a new FILTERS import to views/objects.py and modified the code to use that imported
filter list.