improv: Add entrypoint for out-of-tree backends
This will make it straightforward to replace any use of out-of-tree
backends using the dotted module path feature which was deprecated in a
previous commit. Instead, the out-of-tree author may now simply define
one or more `medallion.backends` entrypoints which can then be
instantiated from  a medallion configuration by their entrypoint name.