memory backend: add a message when an object was requested to be added to a collection, but the collection already had the object. The message just says the object was already added.
mongo backend: remove the verbose success message which was used
when the add operation completed normally.  It seems verbose and
unnecessary (the message property in the status-details object
is optional).  Instead, don't include a message at all if the
add operation completed normally; include a message when the
collection already contained the object.  The message just says
the object was already added.